fixed broken format string modifiers for size_t typed arguments

+++ lib/abyss/src/socket_unix.c
@@ -193,8 +193,8 @@ channelWrite(TChannel *            const channelP,
                 fprintf(stderr, "Abyss channel: send() failed.  "
                         "Socket closed.\n");
             else
-                fprintf(stderr, "Abyss channel: sent %u bytes: '%.*s'\n",
-                        rc, rc, &buffer[len-bytesLeft]);
+                fprintf(stderr, "Abyss channel: sent %zu bytes: '%.*s'\n",
+                        rc, (int)rc, &buffer[len-bytesLeft]);
         }
         if (rc <= 0)
             /* 0 means connection closed; < 0 means severe error */
